+++ core/execution/engine.go
@@ -283,8 +283,8 @@ func (e *Engine) IsEligibleForProposerBonus(marketID string, value *num.Uint) bo
 }
 
 // SubmitMarket submits a new market configuration to the network.
-func (e *Engine) SubmitMarket(ctx context.Context, marketConfig *types.Market, proposer string) error {
-	return e.submitOrRestoreMarket(ctx, marketConfig, proposer, true)
+func (e *Engine) SubmitMarket(ctx context.Context, marketConfig *types.Market, proposer string, oos time.Time) error {
+	return e.submitOrRestoreMarket(ctx, marketConfig, proposer, true, oos)
 }
 
 // RestoreMarket restores a new market from proposal checkpoint.
@@ -293,10 +293,11 @@ func (e *Engine) RestoreMarket(ctx context.Context, marketConfig *types.Market) 
 	if len(proposer) == 0 {
 		return ErrMarketDoesNotExist
 	}
-	return e.submitOrRestoreMarket(ctx, marketConfig, "", false)
+
+	return e.submitOrRestoreMarket(ctx, marketConfig, "", false, e.timeService.GetTimeNow())
 }
 
-func (e *Engine) submitOrRestoreMarket(ctx context.Context, marketConfig *types.Market, proposer string, isNewMarket bool) error {
+func (e *Engine) submitOrRestoreMarket(ctx context.Context, marketConfig *types.Market, proposer string, isNewMarket bool, oos time.Time) error {
 	if e.log.IsDebug() {
 		msg := "submit market"
 		if !isNewMarket {
@@ -305,7 +306,7 @@ func (e *Engine) submitOrRestoreMarket(ctx context.Context, marketConfig *types.
 		e.log.Debug(msg, logging.Market(*marketConfig))
 	}
 
-	if err := e.submitMarket(ctx, marketConfig); err != nil {
+	if err := e.submitMarket(ctx, marketConfig, oos); err != nil {
 		return err
 	}
 
@@ -353,12 +354,10 @@ func (e *Engine) publishUpdateMarketInfos(ctx context.Context, mkt *Market) {
 }
 
 // submitMarket will submit a new market configuration to the network.
-func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market) error {
+func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market, oos time.Time) error {
 	if len(marketConfig.ID) == 0 {
 		return ErrNoMarketID
 	}
-
-	now := e.timeService.GetTimeNow()
 
 	// ensure the asset for this new market exists
 	asset, err := marketConfig.GetAsset()
@@ -372,7 +371,7 @@ func (e *Engine) submitMarket(ctx context.Context, marketConfig *types.Market) e
 	}
 
 	// create market auction state
-	mas := monitor.NewAuctionState(marketConfig, now)
+	mas := monitor.NewAuctionState(marketConfig, oos)
 	ad, err := e.assets.Get(asset)
 	if err != nil {
 		e.log.Error("Failed to create a new market, unknown asset",