+++ core/governance/proposal.go
@@ -143,11 +143,16 @@ func (t *ToSubmit) NewMarket() *ToSubmitNewMarket {
 }
 
 type ToSubmitNewMarket struct {
-	m *types.Market
+	m   *types.Market
+	oos time.Time // opening auction start
 }
 
 func (t *ToSubmitNewMarket) Market() *types.Market {
 	return t.m
+}
+
+func (t *ToSubmitNewMarket) OpeningAuctionStart() time.Time {
+	return t.oos
 }
 
 type VoteClosed struct {