+++ core/integration/steps/execution.go
@@ -14,6 +14,7 @@ package steps
 
 import (
 	"context"
+	"time"
 
 	"code.vegaprotocol.io/vega/core/types"
 )
@@ -30,7 +31,7 @@ type Execution interface {
 		deterministicID string) error
 	AmendLiquidityProvision(ctx context.Context, amendment *types.LiquidityProvisionAmendment, party string) error
 	CancelLiquidityProvision(ctx context.Context, cancel *types.LiquidityProvisionCancellation, party string) error
-	SubmitMarket(ctx context.Context, marketConfig *types.Market, proposer string) error
+	SubmitMarket(ctx context.Context, marketConfig *types.Market, proposer string, oos time.Time) error
 	StartOpeningAuction(ctx context.Context, marketID string) error
 	UpdateMarket(ctx context.Context, marketConfig *types.Market) error
 	BlockEnd(ctx context.Context)