+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [8226](https://github.com/vegaprotocol/vega/issues/8226) - Fix auto initialise failure when initialising empty node
 - [8353](https://github.com/vegaprotocol/vega/issues/8353) - Improve ledger entry `CSV` export.
 - [8358](https://github.com/vegaprotocol/vega/issues/8358) - Fix replay protection
+- [8451](https://github.com/vegaprotocol/vega/issues/8451) - Fix invalid auction duration for new market proposals.
 
 ## 0.71.5